// src/interfaces/Pokemon.ts
export interface ResponseAPI {
  message: string;
  myPokemon: MyPokemon;
}

export interface MyPokemon {
  id: number;
  name: string;
  sprites: Sprites;
  types: string[];
}

export interface Sprites {
  back_default: string;
  back_female: string;
  back_shiny: string;
  back_shiny_female: string;
  front_default: string;
  front_female: string;
  front_shiny: string;
  front_shiny_female: string;
}